fix(nixos): fix emacs service not starting
This was being caused by the emacsql-sqlite binary being outdated
after upgrading (doom)-emacs. As a result, it was missing for the
current version, but also not being generated dynamically, like most
binary versions of packages are.
Hopefully, including the package inside of the home manager
configuration will fix this issue.