hack(qtile): fix emacsclient not working correctly
This appears to have been caused by auto-starting multiple emacs
instances in the scratchpads, in addition to the emacs server started
through systemd. As a result, the latter would not start correctly,
leading to broken instances of emacsclient.
As a workaround, wait two seconds before auto-starting the contents in
the scratchpads.